checkstyling

+++ src/main/java/core/basesyntax/service/impl/PurchaseAction.java
@@ -6,6 +6,9 @@ import core.basesyntax.service.ActionHandler;
 public class PurchaseAction implements ActionHandler {
     @Override
     public void apply(String fruit, int quantity) {
+        if (quantity < 0) {
+            throw new RuntimeException("Quantity can't be negative");
+        }
         Storage.removeFruit(fruit, quantity);
     }
 }
